feat(participation): 참여율 공통 템플릿 - 기간을 시트 안에서 입력받아 스스로 세팅
사업별 시트 생성기(#643)를 걷어내고 공통 양식 하나로 바꾼다. 사업마다 발급하면 발급
자체가 운영 부담이다. cashflow 처럼 공통 양식을 복사해 쓰고, 기간은 시트 맨 위에서
시작월·종료월을 고르면 월 헤더가 수식으로 스스로 생긴다(120열 미리 깔림, 기간 밖은
빈 헤더+회색). 과거 값 프리필도 걷어냈다 - 과거는 리셋하고 새로 받는다는 결정.

- 월 헤더(2행)가 수식: 첫 칸=시작월, 다음 칸=EDATE+1, 종료월에서 멈춤. 파서는 이 행만
  읽는다. 1행 연도 표시는 병합 없이 수식으로만(연도가 바뀌는 첫 달에 표시).
- 투입시작·종료월 드롭다운의 출처는 월 헤더 행 자신 - 고른 기간 밖은 목록에 없다.
- 기간 설정 칸과 투입월 열은 텍스트 서식(@) 고정. 구글 시트가 "2025-07" 입력을 날짜로
  변환해 텍스트 목록과 달라지면서 드롭다운으로 고른 값조차 거부되는 실제 버그를 막는다.
- 기본투입률 자동 채움·diff 노랑·미입력 빨강·닉네임 드롭다운 강제는 그대로.

반영(불러오기) 때 시트 기간과 플랫폼 계약 기간을 대조한다. 시트 입력은 편의이고 진실
검증은 플랫폼 몫이다.
